+++ R/filter_deployments.R
@@ -6,8 +6,8 @@
 #'
 #' - Media are filtered on associated `deploymentID`.
 #' - Observations are filtered on associated `deploymentID`.
-#' - Metadata (`x$temporal`, `x$spatial` and `x$taxonomic`) are updated to
-#' match the filtered deployments.
+#' - Metadata (`x$spatial`, `x$temporal` and `x$taxonomic`) are updated to match
+#' the filtered deployments.
 #'
 #' @inheritParams print.camtrapdp
 #' @param ... Filtering conditions, see `dplyr::filter()`.
@@ -58,19 +58,10 @@ filter_deployments <- function(x, ...) {
     observations(x) %>%
     dplyr::filter(.data$deploymentID %in% deployments$deploymentID)
 
-  # Assign filtered data
+  # Assign filtered data (also updates spatial, temporal and taxonomic scope)
   deployments(x) <- deployments
   media(x) <- media
   observations(x) <- observations
-
-  # Update temporal and spatial scope in metadata
-  x <-
-    x %>%
-    update_temporal() %>%
-    update_spatial()
-
-  # Update taxonomic scope in metadata
-  x <- update_taxonomic(x)
 
   return(x)
 }