+++ tests/testthat/test-taxa.R
@@ -27,3 +27,100 @@ test_that("taxa() returns the expected rows (unique taxa)", {
     nrow(observations(x))
   )
 })
+
+test_that("taxa() removes duplicates and keeps the taxon with most taxonomical
+          information", {
+  skip_if_offline()
+  x <- example_dataset()
+
+  # Duplicate sci names: keep first record with taxonID
+  observations(x) <- data.frame(
+    observationID = c(
+      "delete:no_taxonID",
+      "keep",
+      "delete:exact_duplicate",
+      "delete:not_first",
+      "delete:not_first",
+      "keep:diff_name",
+      "delete:not_first"
+    ),
+    scientificName = c(
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas strepera",
+      "Anas strepera"
+    ),
+    taxon.taxonID = c(
+      NA_character_,
+      "DGP6",
+      "DGP6",
+      "DGP6_1",
+      "DGP6_2",
+      "DGPL",
+      "DGPL_1"
+    )
+  )
+  expected_taxa <- dplyr::tibble(
+    scientificName = c("Anas platyrhynchos", "Anas strepera"),
+    taxonID = c("DGP6", "DGPL")
+  )
+  expect_identical(taxa(x), expected_taxa)
+
+  # Duplicate with different amount of info: keep with most
+  observations(x) <- data.frame(
+    observationID = c(
+      "delete:1_columns",
+      "delete:4_columns",
+      "keep:5_columns"
+    ),
+    scientificName = c(
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas platyrhynchos"
+    ),
+    taxon.taxonRank = c(
+      NA_character_,
+      "species",
+      "species"
+    ),
+    taxon.class = c(
+      NA_character_,
+      "Animalia",
+      NA_character_
+    ),
+    taxon.family = c(
+      NA_character_,
+      NA_character_,
+      "Anatidae"
+    ),
+    taxon.vernacularNames.eng = c(
+      NA_character_,
+      "mallard",
+      "mallard"
+    ),
+    taxon.vernacularNames.fra = c(
+      NA_character_,
+      NA_character_,
+      NA_character_
+    ),
+    taxon.vernacularNames.nld = c(
+      NA_character_,
+      NA_character_,
+      "wilde eend"
+    )
+  )
+  expected_taxa <- dplyr::tibble(
+    scientificName = c("Anas platyrhynchos"),
+    taxonRank = c("species"),
+    # class: removed because duplicate with less columns has it
+    family = c("Anatidae"),
+    vernacularNames.eng = c("mallard"),
+    # vernacularNames.fra removed because no data
+    vernacularNames.nld = c("wilde eend")
+  )
+  expect_identical(taxa(x), expected_taxa)
+})
+